chore(cmake): moved Gmp detection upwards to pick dynamic linking
We used to prefer static linking for seahorn binaries. For that reason,
we have customized cmake to find static libraries to link against whenever
possible.

This has been cause of many linking issues. In the past, the motivation was that
statically linked binaries are easier to distribute to others. However, our
current distribution mode is a docker container that solves the dependency
problem in a different way.

We should probably move to dynamic linking throughout, but for now, I've just
moved Gmp since it created some strange linking error on Ubuntu bionic with
clang-15.