fix(z2expr): conversion of arrays
In SMT-LIB and Z3, const-array takes two arguments: array sort and a value.

In Seahorn/Expr, const-array takes two arguments: domain sort and a value.

The converter did not properly convert between the two.

The Z3 version is better since it also works with multi-dimensional arrays.
Seahorn version is simpler for single-dimensional arrays.

We should port seahorn to use the standard representation of const-array