fix(dummymainpass): run only once since pass is not idempotent
Earlier this pass used to add a main if one was absent.
Now we are more aggressive and replace main.
This means that if we run more than once then we will undo the work of first run in second run.
Add to this, that inlining runs inbetween the passes -- we end up with a main that is a NOP.