fix: bug #558
Command `smc` was registered twice in the argparse in sea script.
Once as command for SimpleMemoryChecker, and another as a compound
command that runs SimpleMemoryChecker as part of the pipeline.

Before python3.11, the new command was silently overriding the old.
At python3.11, argparse module complains about the conflict.

Renamed smc internal command to `smc-check`, and left `smc` as the
name for the compound command.